lustre: remove obsolete OBD_FAIL_OSC_DIO_PAUSE fail_loc
The fail_loc used for testing was removed in Lustre 2.0.
The fail_loc tests for a bug which should be obvious - a
serious delay when doing DIO writes - and is definitely
fixed in current versions.  (Bugzilla 15950)

And without the fail_loc, the test isn't doing anything
interesting.  But the timer based aspect of it fails
occasionally due to hardware delays.  So let's just remove
the fail_loc.

WC-bug-id: https://jira.whamcloud.com/browse/LU-13706
lustre-commit: 59d5bb1558b281d75 ("LU-13706 tests: remove test 119d")
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50731
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>